move wallet connect types to types package
