+++ packages/renderer/src/renderer/contexts/import/WalletConnect/types.ts
@@ -1,8 +1,10 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { AnyData } from '@polkadot-live/types/misc';
-import type { ChainID } from '@polkadot-live/types/chains';
+import type {
+  WcFetchedAddress,
+  WcSelectNetwork,
+} from '@polkadot-live/types/walletConnect';
 
 export interface WalletConnectImportContextInterface {
   wcFetchedAddresses: WcFetchedAddress[];
@@ -14,21 +16,4 @@ export interface WalletConnectImportContextInterface {
   setWcFetchedAddresses: React.Dispatch<
     React.SetStateAction<WcFetchedAddress[]>
   >;
-}
-
-/**
- * @todo Move to types package or import window
- */
-export interface WcSelectNetwork {
-  caipId: string;
-  ChainIcon: AnyData;
-  chainId: ChainID;
-  selected: boolean;
-}
-
-export interface WcFetchedAddress {
-  chainId: ChainID;
-  encoded: string;
-  substrate: string;
-  selected: boolean;
 }
